Fixes #15
Not all that unusual, that comment conflicts with the installation guide (Zenoss_Core_Installation_Upgrade_01-082013-4.2-v15.pdf) which stats "Zenoss Core requires at least MySQL 5.5.25, and supports later versions through 5.5.31".

I've justed tested using 5.5.36 (latest available) on EL6 with 4.2.5 and it seems to be
working fine. Going to roll forward with 5.5.36 unless there is a compelling reason not to.

Also adding MySQL-shared-compat package